AJAX frontend for problem page
This prevents re-loading of the side when running the code on the
problem page, it uses an ajax-request instead.

The sandbar.stateful-session is misused a bit in order to minimally
affect existing code. An alternative would be to make "run-code",
"mark-completed" and some other functions more functional and move all
the side effects (if needed) into "static-run-code" and "rest-run-code".

The animation on the frontend while the code is executing is not too
pretty, it is just meant as a start.