Slow down the AJAX interface a little bit
When pressing the "Run" Button, the unit tests are executed a little
too fast with little visual clues, which can have disorienting effects.

This makes the green/red light(s) light up with 500ms delays, one after
the other. The updated Message is displayed when the light of last test
xor the failing test is updated.

Please see the discussion here:
https://github.com/dbyrne/4clojure/pull/83