Add width/height attributes to images on front page.
This means the page will render correctly on the first try;
if you don't, it has to guess an image size, render stuff,
and then move it all around when it figures out the size.

If anyone knows a better way to set image sizes without
hard-coding them into the source code, I'd love to hear it.