Allow people with current accounts to associate an OpenID
This is a rather large update that allows users with existing accounts
to associate an OpenID with their current account.  When a user logs
in, the :user returned by session/get-session is now a map, but this
is made backwards-compatible by most of the utility functions
accepting either a string or a map.